picaxeが作り直し

+++ 2.5D BaseFramework+ImGui/Src/Application/System/Timer.cpp
@@ -2,7 +2,8 @@
 
 void Timer::Start() noexcept
 {
-    if (!m_isRunning) {
+    if (!m_isRunning)
+	{
         auto now         = std::chrono::steady_clock::now();
         if (m_startTime != std::chrono::steady_clock::time_point()) m_startTime += (now - m_stopTime);
         else m_startTime = now;
@@ -12,7 +13,8 @@ void Timer::Start() noexcept
 
 void Timer::Stop() noexcept
 {
-    if (m_isRunning) {
+    if (m_isRunning)
+	{
         m_stopTime  = std::chrono::steady_clock::now();
         m_isRunning = false;
     }
@@ -27,13 +29,14 @@ void Timer::Restart() noexcept
 
 void Timer::Resume() noexcept
 {
-    if (!m_isRunning) {
+    if (!m_isRunning) 
+	{
         m_startTime += (std::chrono::steady_clock::now() - m_stopTime);
         m_isRunning  = true;
     }
 }
 
-size_t const Timer::ElapsedSeconds() const noexcept
+long long const Timer::ElapsedSeconds() const noexcept
 {
 	auto endTime = m_isRunning ? std::chrono::steady_clock::now() : m_stopTime;
 	return std::chrono::duration_cast<std::chrono::seconds>(endTime - m_startTime).count();
